requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.11
h11==0.14.0
idna==3.10
imbalanced-learn==0.13.0
joblib==1.4.2
numpy==2.2.4
nvidia-nccl-cu12==2.26.2
pandas==2.2.3
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
starlette==0.46.1
threadpoolctl==3.6.0
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
xgboost==3.0.0
